upadet Lab3 - fix command error
